+++ src/main/java/nextstep/ladder/domain/Moveable.java
@@ -1,0 +1,6 @@
+package nextstep.ladder.domain;
+
+public interface Moveable {
+
+    Position moveOn(Position from);
+}
